Provide Web API test runner that conforms to meta harness requirements
Introduces program webapitests which conforms to the meta harness'
required flag layout.  This can be used to list logical test groups, all tests,
and run filtered test runs.

It currently only includes semiauto.TestCase based tests, which is
a known defect.  This can be rectified in the future by making the
browser-based test runner available to the MinimalTestCase based
tests.

Another known issue is that it doesn't (yet) couple the whole thing
up to the meta harness.  This will be addressed in a future patch.