Adding SECCOMP status check
FxOS versions 2.0+ have Seccomp support, or at least they should have.
This commit adds the 'kernel' test group with a single seccomp' test.
It runs b2g-ps through adb and checks its output to determine if the
b2g process is running with seccomp enabled, and it FAILs the test if
it doesn't.